Refactor parquet fs functions
